move date format constants to single file
