+++ components/pages/Applications/ApplicationForm/RequestRevisionsBar/PDFActions.tsx
@@ -30,7 +30,7 @@ import { ApprovedDoc } from '../../types';
 import { CustomLoadingButton } from '../Forms/common';
 import { DOCUMENT_TYPES } from '../Forms/types';
 import { format as formatDate } from 'date-fns';
-import { API_DEFAULT_DATE_FORMAT } from '../../Dashboard/Applications/InProgress/constants';
+import { API_DEFAULT_DATE_FORMAT } from 'global/utils/dates/constants';
 import { SetLastUpdated } from '../types';
 import { css } from '@icgc-argo/uikit';
 import Typography from '@icgc-argo/uikit/Typography';
