+++ global/constants/index.ts
@@ -26,10 +26,6 @@ export const EGO_JWT_KEY = 'egoJwt';
 export const ADMIN_APPLICATIONS_LABEL = 'Manage Applications';
 export const APPLICANT_APPLICATIONS_LABEL = 'My Applications';
 
-export const DATE_RANGE_DISPLAY_FORMAT = 'Y-MM-dd';
-
-export const DATE_TEXT_FORMAT = 'MMM. dd, yyyy';
-
 export const SUBMISSION_SUCCESS_CHECK = 'daco.submission-success-check';
 export const APPROVED_APP_CLOSED_CHECK = 'daco.approved-app-closed-check';
 export const GDPR_ACCEPTED = 'gdpr-accepted';
